challenge-05-qr-and-codes/src/requirements.txt
pillow
qrcode
